+++ src/onvif_media_snapshot.h
@@ -5,8 +5,8 @@
 
 typedef struct _OnvifSnapshot OnvifSnapshot;
 
-SHARD_EXPORT OnvifSnapshot * OnvifSnapshot__create(int size, char * buffer);
-SHARD_EXPORT void OnvifSnapshot__init(OnvifSnapshot * self, int size, char * buffer);
+SHARD_EXPORT OnvifSnapshot * OnvifSnapshot__create(size_t size, char * buffer);
+SHARD_EXPORT void OnvifSnapshot__init(OnvifSnapshot * self, size_t size, char * buffer);
 SHARD_EXPORT void OnvifSnapshot__destroy(OnvifSnapshot * self);
 
 SHARD_EXPORT double OnvifSnapshot__get_size(OnvifSnapshot * self);
